criacao do cargo e exception
