--- src/main/java/br/com/ciss/crud_funcionario/dtos/FuncionarioDto.java
@@ -1,62 +1,0 @@
-package br.com.ciss.crud_funcionario.dtos;
-
-import jakarta.validation.constraints.Size;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-public class FuncionarioDto {
-
-    private Long id;
-
-    @Size(min = 3, max = 50)
-    private String nome;
-
-    @Size(min = 2, max = 50)
-    private String sobrenome;
-
-    private String email;
-
-    private String nis;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getSobrenome() {
-        return sobrenome;
-    }
-
-    public void setSobrenome(String sobrenome) {
-        this.sobrenome = sobrenome;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getNis() {
-        return nis;
-    }
-
-    public void setNis(String nis) {
-        this.nis = nis;
-    }
-}
