+++ src/main/java/br/com/ciss/crud_funcionario/funcionario/Cargo.java
@@ -1,0 +1,15 @@
+package br.com.ciss.crud_funcionario.funcionario;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Cargo {
+
+    private String cargo;
+}
