+++ src/main/java/br/com/ciss/crud_funcionario/funcionario/Funcionario.java
@@ -1,0 +1,35 @@
+package br.com.ciss.crud_funcionario.funcionario;
+
+import jakarta.persistence.*;
+import jakarta.validation.constraints.Size;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "funcionario")
+public class Funcionario {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Size(min = 3, max = 50)
+    private String nome;
+
+    @Size(min = 2, max = 50)
+    private String sobrenome;
+
+    private String email;
+
+    private String nis;
+
+    private String cargo;
+
+
+}
